Split `syntax_mapping/builtin.rs`
